Fix six file-format bugs found by the dalib-ts 7.41 reconciliation
Angelique's TypeScript port (dalib-ts, agent Fintan) surfaced these while
reconciling against a real Dark Ages 7.41 client and the darkages-741-re
docs. Each was re-verified here against the RE docs and, where possible,
against real retail assets before fixing.

1. ControlFileParser: <IMAGE> is an explicit ordered list of (name, frame)
   visual states, not a start/end range. Range-filling fabricated frames for
   sparse lists. Confirmed live on retail data: field001.txt has
   emot000.epf {0,1,3} and emot001.epf {0,1,7}, which the old code expanded
   to {0,1,2,3} and {0..7}, shifting every later emote button state.

2. Graphics.RenderTile: ground tiles are isometric diamonds. Inside the
   diamond palette index 0 is an ordinary opaque color (not transparent as
   for sprites), and pixels outside the diamond are masked transparent. The
   old path routed tiles through the sprite SimpleRender, dropping index-0
   ground pixels (holes) and applying no mask.

3. MapFile: cells are u16le. A foreground id above 32767 (stc ids are five
   digits) read as a negative short and was then treated as non-renderable.
   Also, the client accepts a longer-than-expected file and ignores the
   trailing bytes; the old strict-equality length check rejected it.
   MapTile id fields widened short -> ushort (breaking; wire-faithful).

4. HeaFile: a run's intensity byte carries flags in its top two bits; mask
   with & 0x3F at decode. MAX_LIGHT_VALUE is 0x20, so a flag bit corrupted
   the intensity.

5. SpfFile/SpfView: frame size is the visible rectangle (Right-Left x
   Bottom-Top), and source rows advance by ByteWidth (pitch). The old decode
   used absolute Right/Bottom and ignored pitch, so nonzero-origin or padded
   frames decoded wrong. No shipped 7.41 asset has such a frame, so this is
   robustness verified by synthetic tests.

6. PaletteTable: strip // comments and split on whitespace runs in both the
   range parser and ParseCyclingFile. The client skips blank lines and //
   comments; the old single-space split dropped commented/padded entries.

Adds 11 regression tests (945 total, all passing).

+++ DALib/Extensions/IntExtensions.cs
@@ -25,5 +25,5 @@ public static class IntExtensions
     ///     0-12 and 10000-10012 are not rendered by the client... 20000-20012 are rendered but are kinda buggy if you get
     ///     close to them
     /// </remarks>
-    public static bool IsRenderedTileIndex(this short tileIndex) => (tileIndex > 10012) || ((tileIndex % 10000) > 12);
+    public static bool IsRenderedTileIndex(this int tileIndex) => (tileIndex > 10012) || ((tileIndex % 10000) > 12);
 }
